+++ lib/apply/diff.js
@@ -16,15 +16,15 @@ var _prepareConnectionForDiff = require('../transformers/connections/prepareConn
 
 var _prepareConnectionForDiff2 = _interopRequireDefault(_prepareConnectionForDiff);
 
-var _additions = require('./diff/additions');
+var _additions = require('./diffFunctions/additions');
 
 var _additions2 = _interopRequireDefault(_additions);
 
-var _removals = require('./diff/removals');
+var _removals = require('./diffFunctions/removals');
 
 var _removals2 = _interopRequireDefault(_removals);
 
-var _changes = require('./diff/changes');
+var _changes = require('./diffFunctions/changes');
 
 var _changes2 = _interopRequireDefault(_changes);
