+++ lib/apply/diffFunctions/additions.js
