+++ lib/apply/diffFunctions/changes.js
@@ -23,7 +23,7 @@ function selectiveEquals(filter) {
 }
 
 var deepMerge = _ramda2.default.mergeWith(function (l, r) {
-  return _ramda2.default.isEmpty(r) ? r : _ramda2.default.is(Array, l) && _ramda2.default.is(Array, r) ? _ramda2.default.concat(l, r) : !(_ramda2.default.is(Object, l) && _ramda2.default.is(Object, r)) ? r : deepMerge(l, r);
+  return _ramda2.default.isEmpty(r) || _ramda2.default.is(Array, l) && _ramda2.default.is(Array, r) || !(_ramda2.default.is(Object, l) && _ramda2.default.is(Object, r)) ? r : deepMerge(l, r);
 });
 
 var bothExist = function bothExist(lhs, rhs) {
