+++ lib/apply/diffFunctions/removals.js
