+++ lib/apply/index.js
@@ -9,10 +9,6 @@ var _ramda = require('ramda');
 
 var _ramda2 = _interopRequireDefault(_ramda);
 
-var _loadEnv = require('../utils/loadEnv');
-
-var _loadEnv2 = _interopRequireDefault(_loadEnv);
-
 var _say2 = require('../utils/say');
 
 var _say3 = _interopRequireDefault(_say2);
@@ -21,27 +17,9 @@ var _inquirer = require('inquirer');
 
 var _inquirer2 = _interopRequireDefault(_inquirer);
 
-var _loadManifest = require('../apply/loadManifest');
+var _prepareAll = require('../utils/prepareAll');
 
-var _loadManifest2 = _interopRequireDefault(_loadManifest);
-
-var _setupClient = require('../utils/setupClient');
-
-var _setupClient2 = _interopRequireDefault(_setupClient);
-
-var _fetchRules = require('../auth0/rules/fetchRules');
-
-var _fetchRules2 = _interopRequireDefault(_fetchRules);
-
-var _fetchConnections = require('../auth0/connections/fetchConnections');
-
-var _fetchConnections2 = _interopRequireDefault(_fetchConnections);
-
-var _fetchClients = require('../auth0/clients/fetchClients');
-
-var _fetchClients2 = _interopRequireDefault(_fetchClients);
-
-var _associateClientConnections = require('../utils/associateClientConnections');
+var _prepareAll2 = _interopRequireDefault(_prepareAll);
 
 var _diff = require('./diff');
 
@@ -51,9 +29,9 @@ var _applyDiff = require('./applyDiff');
 
 var _applyDiff2 = _interopRequireDefault(_applyDiff);
 
-var _printDiff = require('./printDiff');
+var _printApply = require('./printApply');
 
-var _printDiff2 = _interopRequireDefault(_printDiff);
+var _printApply2 = _interopRequireDefault(_printApply);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -88,6 +66,6 @@ function index() {
   var error = _say.error;
 
 
-  return (0, _loadEnv2.default)(options).then(ensure).then((0, _loadManifest2.default)(filename)).then(_setupClient2.default).then(_fetchRules2.default).then(_fetchConnections2.default).then(_fetchClients2.default).then(_associateClientConnections.disassociate).then((0, _diff2.default)('rules')).then((0, _diff2.default)('connections')).then(_ramda2.default.ifElse(isDryRun, _printDiff2.default, _applyDiff2.default)).catch(error);
+  return (0, _prepareAll2.default)(filename, options).then(ensure).then((0, _diff2.default)('rules')).then((0, _diff2.default)('connections')).then(_ramda2.default.ifElse(isDryRun, _printApply2.default, _applyDiff2.default)).catch(error);
 }
 //# sourceMappingURL=index.js.map
