+++ lib/cli.js
@@ -61,7 +61,7 @@ function addDumpCommand(cli) {
 }
 
 function addApplyCommand(cli) {
-  cli.command('apply [rules file]').option('-e, --env <env>', "Auth0 env to run against").option('-k, --key <key>', "The Auth0 domain key.").option('-s, --secret <secret>', "The Auth0 domain secret.").option('-d, --dry-run', "Perform a dry run, outputting the diff that will be applied.").option('-y, --yes', "Bypass confirmation step.").action(_index8.default);
+  cli.command('apply [rules file]').option('-e, --env <env>', "Auth0 env to run against").option('-k, --key <key>', "The Auth0 domain key.").option('-s, --secret <secret>', "The Auth0 domain secret.").option('-d, --dry-run', "Perform a dry run, outputting the config that will be applied.").option('-y, --yes', "Bypass confirmation step.").action(_index8.default);
 
   return cli;
 }
